edit,delete creating

+++ Spring_project/OrlCrew/src/main/java/com/orl/crew/service/CrewManageService.java
@@ -1,11 +1,17 @@
 package com.orl.crew.service;
 
+import java.io.File;
+
+import javax.servlet.http.HttpServletRequest;
+
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.orl.crew.dao.Dao;
 import com.orl.crew.domain.Crew;
+import com.orl.crew.domain.CrewInsertRequest;
+import com.orl.member.domain.Member;
 
 @Service
 public class CrewManageService {
@@ -15,6 +21,25 @@ public class CrewManageService {
 	@Autowired
 	SqlSessionTemplate template;
 	
+	public boolean isHaveAuth(
+			int crewIdx,
+			HttpServletRequest request
+			) {
+		boolean chk = false;
+		Member member = (Member)request.getSession().getAttribute("member");
+		int nowAuthIdx = member.getMemberIdx();
+		int crewAuthIdx = selectCrew(crewIdx).getMemberIdx();
+		
+		try {
+			if(nowAuthIdx == crewAuthIdx) {
+				chk = true;
+			}
+		} catch (NullPointerException e) {
+			e.printStackTrace();
+		}
+		return chk;
+	}
+	
 	public Crew selectCrew(
 			int crewIdx
 			) {
@@ -23,16 +48,50 @@ public class CrewManageService {
 		return crew;
 	}
 	
+	public int updateCrew(
+			CrewInsertRequest crewRequest,
+			HttpServletRequest request
+			) {
+		int resultCnt = 0;
+		File newFile = null;
+		dao = template.getMapper(Dao.class);
+		
+		CrewInsertService insertservice = new CrewInsertService();
+		
+		try {
+			if (crewRequest.getCrewPhoto() != null && !crewRequest.getCrewPhoto().isEmpty()) {
+				newFile = insertservice.saveFile(request, crewRequest.getCrewPhoto());
+				resultCnt = dao.updateCrew(crewRequest.getCrewName(), 
+						newFile.getName(), 
+						crewRequest.getCrewDiscription(), 
+						crewRequest.getCrewTag());
+			} else {
+				resultCnt = dao.updateCrewWithoutPhoto(crewRequest.getCrewName(), 
+						crewRequest.getCrewDiscription(), 
+						crewRequest.getCrewTag());
+			}
+		} catch(Exception e) {
+			e.printStackTrace();
+			if(newFile != null && newFile.exists()) {
+				newFile.delete();
+				System.out.println("오류가 발생하여 파일을 삭제했습니다.");
+			}
+		}
+		return resultCnt;
+	}
+	
 	public int deleteCrew(
 			int crewIdx,
 			String crewName
 			) {
-		
 		int resultCnt = 0;
-		
 		dao = template.getMapper(Dao.class);
-		resultCnt = dao.deleteCrew(crewIdx);
 		
+		Crew crew = selectCrew(crewIdx);
+		if(crew.getCrewName().contentEquals(crewName)) {
+			resultCnt = dao.deleteCrew(crewIdx);
+
+		}
 		return resultCnt;
 	}
 }
